remove the `query_filter` module
