+++ src/tests.rs
@@ -1,5 +1,4 @@
 use super::*;
-use crate::query_filter::{OneAdded, OneChanged};
 use std::fmt::{Debug, Display};
 
 #[derive(Resource, Default)]
